[BACKLOG-18865]-move getting parameters for parquet step to base meta, to receive them in spark in the same way - environment substitute, variable space, parse integers
