+++ kettle-plugins/formats-meta/src/main/resources/org/pentaho/big/data/kettle/plugins/formats/parquet/output/messages/messages_en_US.properties
@@ -1,0 +1,7 @@
+ParquetOutput.EncodingType.PLAIN=Plain
+ParquetOutput.EncodingType.DICTIONARY=Dictionary
+ParquetOutput.EncodingType.BIT_PACKED=Bit packed
+ParquetOutput.EncodingType.RLE=RLE
+
+ParquetOutput.CompressionType.NONE=None
+
