+++ kettle-plugins/formats/src/main/resources/org/pentaho/big/data/kettle/plugins/formats/impl/parquet/output/messages/messages_en_US.properties
@@ -18,10 +18,3 @@ ParquetOutputDialog.Options.RowSize=Row group size (MB):
 ParquetOutputDialog.Options.PageSize=Data page size (KB):
 ParquetOutputDialog.Options.Encoding=Encoding:
 ParquetOutputDialog.Options.DictPageSize=Dictionary page size (KB):
-
-ParquetOutput.EncodingType.PLAIN=Plain
-ParquetOutput.EncodingType.DICTIONARY=Dictionary
-ParquetOutput.EncodingType.BIT_PACKED=Bit packed
-ParquetOutput.EncodingType.RLE=RLE
-
-ParquetOutput.CompressionType.NONE=None
